accounts/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
dotenv==0.9.9
gunicorn==20.1.0
idna==3.10
mysqlclient==2.2.7
python-dotenv>=1.0.0
requests==2.32.3
sqlparse==0.5.3
structlog==25.2.0
structured_log==1.9
urllib3==2.3.0